Add custom big decimal summary statistics to get transaction statistics. Add transaction repository implementation using timer to schedule events. Add serializer to zonedDateTime of transactions. Add tests for verticle and transaction repository. Exclude embedded Tomcat dependency from spring-boot-starter-web since I'm going to deploy the service using verticles.

+++ src/tests/com/n26/TransactionFactory.java
@@ -12,16 +12,16 @@ public class TransactionFactory {
         throw new AssertionError("Ensuring noninstaintability");
     }
 
-    public static Transaction getValidTransaction(final BigDecimal amount) {
-        return new Transaction(amount, ZonedDateTime.now());
+    public static Transaction getValidTransaction(final long amount) {
+        return new Transaction(BigDecimal.valueOf(amount), ZonedDateTime.now());
     }
 
-    public static Transaction getInvalidTransaction(final BigDecimal amount) {
-        return new Transaction(amount, ZonedDateTime.now().minus(61, ChronoUnit.SECONDS));
+    public static Transaction getInvalidTransaction(final long amount) {
+        return new Transaction(BigDecimal.valueOf(amount), ZonedDateTime.now().minus(61, ChronoUnit.SECONDS));
     }
 
-    public static Transaction getTransaction(final BigDecimal amount, final ZonedDateTime timestamp) {
-        return new Transaction(amount, timestamp);
+    public static Transaction getTransaction(final long amount, final ZonedDateTime timestamp) {
+        return new Transaction(BigDecimal.valueOf(amount), timestamp);
     }
 
 }
